Make * operations particular to Vector w/ Matrix
A few modifications to the given example were necessary:

1. A vector and an adjoint vector are different types in `Julia`

2. Making `_Matrix` subtype of `AbstractMatrix` would conflict with
previous PR review asking for the removal of `getindex` and `size`
operations

The solution of 1 is clear, just fix the type. For 2, the approach of
`LinearAlgebra.jl` is employed, which can be described as follows

- Create a type (such as `UniformScaling`) that isn't itself a matrix

- Create functions that work as matrices constructors, example:

  `UniformScaling(Bool)(3)` returns a 3x3 matrix

- Overwrite ring operations, to make their evaluations faster

Two conflicting instructions, namely:

1. Rename the parent type to `CoDaMatrix`

2. An abstract parent type is unnecessary

were synthesized as follows: remove any operations from the abstract
parent structure and rename it to `CoDaMatrix`.